+++ microfrontends/app-builder-menu/src/components/hooks/useContent.tsx
@@ -1,12 +1,12 @@
 import React, { ReactNode, useContext } from 'react';
 import { useIntl } from 'react-intl';
-import { ContentType, getContent } from '../content';
+import { ContentType, getContent } from '../../content';
 
 const ContentContext = React.createContext({});
 
 export const ContentProvider = ({ children }: { children: ReactNode }) => {
   const intl = useIntl();
-  const content:ContentType  = getContent(intl);
+  const content: ContentType = getContent(intl);
 
   return (
     <ContentContext.Provider value={content}>
