+++ microfrontends/app-builder-menu/src/content.ts
